Refactored get_members and remove_members interface adapters to be consistent with their use cases
